Initial move of v4 and v5 code
